requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
pinecone
pdfplumber
tiktoken
python-dotenv